Add API GET, POST, PUT and Delete for Employee

+++ src/main/java/com/example/swalayan/repository/EmployeeRepository.java
@@ -3,5 +3,8 @@ package com.example.swalayan.repository;
 import com.example.swalayan.model.Employee;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EmployeeRepository extends CrudRepository <Employee, Integer> {
+import java.util.List;
+public interface EmployeeRepository extends CrudRepository <Employee, Long> {
+//    List<Employee> findByEmpName(String empName);
+
 }
